Fix for issue #1536 to support multi-env setups
Gym setups run multiple environment processes. Each environment process attempts to log to the same identical file causing access violations and failure of the academy to finish initialization. This change adds a unique identifier to the log file, while keeping it in the same folder users expect it to be.